chore: Update server main.go to use chi router
The server's main.go file has been updated to use the chi router instead of the default http.ServeMux. This change improves the routing capabilities of the server and allows for more complex routing configurations. Due to the complexity of the project, Chi is a good addition to the routing of the server.